Changed type of size to float, according to specs

+++ kool-core/src/commonMain/kotlin/de/fabmax/kool/util/MsdfFont.kt
@@ -183,7 +183,7 @@ data class MsdfMeta(
 data class MsdfAtlasInfo(
     val type: String,
     val distanceRange: Float,
-    val size: Int,
+    val size: Float,
     val width: Int,
     val height: Int,
     val yOrigin: String
